requirements.txt
pydantic
python-dotenv

requests
cloudscraper
cfscrape
sqlalchemy

python-telegram-bot
apscheduler
pytz

dependency-injector